Add ToString overrides for ExtensionMethodInfo/PropertyInfo
Added a `ToString` method override in the `ExtensionMethodInfo`
class to provide a string representation of the extension
method, including the receiver parameter's type and the
method itself (either `ReceiverMethod` or `DeclaredMethod`).

Added a `ToString` method override in the `ExtensionPropertyInfo`
class to provide a string representation of the extension
property, including the receiver parameter's type and the
property itself (`ReceiverProperty`).

These changes improve debugging and logging by offering
clearer and more informative representations of extension
methods and properties.